feat: Complete Docker implementation for CI/CD environment consistency (EPIC 23)
- Phase 1: Basic configuration
  - Created pyproject.toml files for all Python services
  - Built comprehensive docker-compose.yml with all 7 services
  - Set up environment variable templates

- Phase 2: Development experience
  - Implemented Docker Compose profiles (core, workers, tools, full, test)
  - Created development support scripts
  - Configured hot reload for development

- Phase 3: CI/CD integration
  - Optimized Dockerfiles with multi-stage builds
  - Created GitHub Actions Docker integration workflow
  - Implemented Docker-based testing strategy with Makefile

- Phase 4: Production readiness
  - Added production deployment configuration
  - Integrated security scanning (Trivy, dependency scan, secrets)
  - Created comprehensive operations documentation

Resolves: CI passes locally but fails on GitHub issue
All core services tested and running successfully
