+++ .env.example
@@ -1,0 +1,79 @@
+# Seiji Watch Environment Variables
+# Copy this file to .env and fill in your values
+
+# Database Configuration
+DATABASE_URL=postgresql://diet_user:diet_password@localhost:5432/diet_tracker
+POSTGRES_DB=diet_tracker
+POSTGRES_USER=diet_user
+POSTGRES_PASSWORD=diet_password
+
+# Redis Configuration
+REDIS_URL=redis://localhost:6379
+
+# Airtable Configuration
+AIRTABLE_API_KEY=your_airtable_api_key_here
+AIRTABLE_BASE_ID=your_airtable_base_id_here
+
+# OpenAI Configuration (for Whisper and embeddings)
+OPENAI_API_KEY=your_openai_api_key_here
+
+# SendGrid Configuration (for email notifications)
+SENDGRID_API_KEY=your_sendgrid_api_key_here
+SENDGRID_FROM_EMAIL=noreply@example.com
+
+# Discord Configuration (optional)
+DISCORD_WEBHOOK_URL=your_discord_webhook_url_here
+
+# JWT Configuration
+JWT_SECRET_KEY=your_super_secret_jwt_key_here_change_in_production
+JWT_ALGORITHM=HS256
+JWT_EXPIRATION_HOURS=24
+
+# Application Settings
+ENVIRONMENT=development
+DEBUG=true
+LOG_LEVEL=INFO
+
+# Frontend Configuration
+NEXT_PUBLIC_API_BASE_URL=http://localhost:8000
+NEXT_PUBLIC_APP_NAME="Seiji Watch"
+NEXT_PUBLIC_APP_DESCRIPTION="Diet Issue Tracker"
+
+# Service Ports (for local development)
+API_GATEWAY_PORT=8000
+WEB_FRONTEND_PORT=3000
+DIET_SCRAPER_PORT=8081
+DATA_PROCESSOR_PORT=8082
+STT_WORKER_PORT=8083
+VECTOR_STORE_PORT=8084
+NOTIFICATIONS_WORKER_PORT=8085
+
+# GCP Configuration (for production)
+GCP_PROJECT=your-gcp-project-id
+GCP_REGION=asia-northeast1
+GCS_BUCKET=your-storage-bucket
+
+# Cloud SQL (for production)
+CLOUD_SQL_CONNECTION_NAME=project:region:instance
+
+# Monitoring
+SENTRY_DSN=your_sentry_dsn_here
+NEW_RELIC_LICENSE_KEY=your_new_relic_key_here
+
+# Feature Flags
+ENABLE_STT_PROCESSING=true
+ENABLE_VECTOR_SEARCH=true
+ENABLE_EMAIL_NOTIFICATIONS=true
+ENABLE_DISCORD_NOTIFICATIONS=false
+
+# Rate Limiting
+RATE_LIMIT_REQUESTS_PER_MINUTE=60
+RATE_LIMIT_REQUESTS_PER_HOUR=1000
+
+# Batch Processing
+BATCH_SIZE=100
+MAX_WORKERS=4
+
+# Cache Settings
+CACHE_TTL_SECONDS=3600
+CACHE_MAX_SIZE_MB=100
