+++ Dockerfile.test
@@ -1,0 +1,51 @@
+# ===================================================
+# Test Runner Dockerfile
+# ===================================================
+# Unified test runner for all services
+
+FROM python:3.11-slim
+
+# Install system dependencies
+RUN apt-get update && apt-get install -y \
+    curl \
+    git \
+    build-essential \
+    postgresql-client \
+    redis-tools \
+    netcat-openbsd \
+    && curl -fsSL https://deb.nodesource.com/setup_18.x | bash - \
+    && apt-get install -y nodejs \
+    && apt-get clean \
+    && rm -rf /var/lib/apt/lists/*
+
+# Install Python test dependencies
+RUN pip install --no-cache-dir \
+    pytest==7.4.3 \
+    pytest-cov==4.1.0 \
+    pytest-asyncio==0.21.1 \
+    pytest-timeout==2.2.0 \
+    pytest-xdist==3.5.0 \
+    coverage==7.3.2 \
+    black==23.12.0 \
+    ruff==0.1.8 \
+    mypy==1.7.1
+
+# Install Node.js test dependencies globally
+RUN npm install -g \
+    jest@29.7.0 \
+    @playwright/test@1.40.0 \
+    eslint@8.55.0 \
+    prettier@3.1.1
+
+# Install Playwright browsers
+RUN npx playwright install --with-deps chromium
+
+# Create test user
+RUN useradd -m -u 1000 testrunner
+USER testrunner
+
+# Set working directory
+WORKDIR /workspace
+
+# Default command
+CMD ["bash", "-c", "echo 'Test runner ready. Use scripts/docker/run-all-tests.sh to run tests.'"]
