+++ nginx/nginx.conf
@@ -1,0 +1,189 @@
+user nginx;
+worker_processes auto;
+error_log /var/log/nginx/error.log warn;
+pid /var/run/nginx.pid;
+
+events {
+    worker_connections 2048;
+    use epoll;
+    multi_accept on;
+}
+
+http {
+    include /etc/nginx/mime.types;
+    default_type application/octet-stream;
+
+    # Logging
+    log_format main '$remote_addr - $remote_user [$time_local] "$request" '
+                    '$status $body_bytes_sent "$http_referer" '
+                    '"$http_user_agent" "$http_x_forwarded_for" '
+                    'rt=$request_time uct="$upstream_connect_time" '
+                    'uht="$upstream_header_time" urt="$upstream_response_time"';
+    
+    access_log /var/log/nginx/access.log main;
+
+    # Performance
+    sendfile on;
+    tcp_nopush on;
+    tcp_nodelay on;
+    keepalive_timeout 65;
+    types_hash_max_size 2048;
+    client_max_body_size 100M;
+
+    # Gzip
+    gzip on;
+    gzip_vary on;
+    gzip_min_length 1024;
+    gzip_proxied any;
+    gzip_comp_level 6;
+    gzip_types text/plain text/css text/xml text/javascript 
+               application/json application/javascript application/xml+rss 
+               application/rss+xml application/atom+xml image/svg+xml 
+               text/x-js text/x-cross-domain-policy application/x-font-ttf 
+               application/x-font-opentype application/vnd.ms-fontobject 
+               image/x-icon;
+
+    # Rate Limiting
+    limit_req_zone $binary_remote_addr zone=api:10m rate=10r/s;
+    limit_req_zone $binary_remote_addr zone=web:10m rate=30r/s;
+    limit_conn_zone $binary_remote_addr zone=addr:10m;
+
+    # Cache
+    proxy_cache_path /var/cache/nginx levels=1:2 keys_zone=cache:10m 
+                     max_size=1g inactive=60m use_temp_path=off;
+
+    # Security Headers
+    add_header X-Frame-Options "SAMEORIGIN" always;
+    add_header X-Content-Type-Options "nosniff" always;
+    add_header X-XSS-Protection "1; mode=block" always;
+    add_header Referrer-Policy "strict-origin-when-cross-origin" always;
+
+    # Upstream servers
+    upstream api_backend {
+        least_conn;
+        server api-gateway:8000 max_fails=3 fail_timeout=30s;
+        keepalive 32;
+    }
+
+    upstream web_backend {
+        least_conn;
+        server web-frontend:3000 max_fails=3 fail_timeout=30s;
+        keepalive 32;
+    }
+
+    # Redirect HTTP to HTTPS (when SSL is configured)
+    server {
+        listen 80;
+        server_name _;
+        
+        # Health check endpoint
+        location /health {
+            access_log off;
+            return 200 "healthy\n";
+            add_header Content-Type text/plain;
+        }
+
+        # Metrics endpoint for Prometheus
+        location /metrics {
+            stub_status;
+            access_log off;
+            allow 172.21.0.0/24;  # backend-tier network
+            deny all;
+        }
+
+        location / {
+            return 301 https://$host$request_uri;
+        }
+    }
+
+    # Main HTTPS server (uncomment when SSL certificates are available)
+    # server {
+    #     listen 443 ssl http2;
+    #     server_name seiji-watch.example.com;
+    #
+    #     ssl_certificate /etc/nginx/ssl/cert.pem;
+    #     ssl_certificate_key /etc/nginx/ssl/key.pem;
+    #     ssl_protocols TLSv1.2 TLSv1.3;
+    #     ssl_ciphers HIGH:!aNULL:!MD5;
+    #     ssl_prefer_server_ciphers on;
+    #     ssl_session_cache shared:SSL:10m;
+    #     ssl_session_timeout 10m;
+    #
+    #     # API Gateway
+    #     location /api/ {
+    #         limit_req zone=api burst=20 nodelay;
+    #         limit_conn addr 10;
+    #
+    #         proxy_pass http://api_backend/;
+    #         proxy_http_version 1.1;
+    #         proxy_set_header Upgrade $http_upgrade;
+    #         proxy_set_header Connection "upgrade";
+    #         proxy_set_header Host $host;
+    #         proxy_set_header X-Real-IP $remote_addr;
+    #         proxy_set_header X-Forwarded-For $proxy_add_x_forwarded_for;
+    #         proxy_set_header X-Forwarded-Proto $scheme;
+    #         proxy_connect_timeout 60s;
+    #         proxy_send_timeout 60s;
+    #         proxy_read_timeout 60s;
+    #
+    #         # Cache GET requests
+    #         proxy_cache cache;
+    #         proxy_cache_valid 200 10m;
+    #         proxy_cache_valid 404 1m;
+    #         proxy_cache_use_stale error timeout updating http_500 http_502 http_503 http_504;
+    #         proxy_cache_lock on;
+    #     }
+    #
+    #     # Web Frontend
+    #     location / {
+    #         limit_req zone=web burst=50 nodelay;
+    #         limit_conn addr 20;
+    #
+    #         proxy_pass http://web_backend;
+    #         proxy_http_version 1.1;
+    #         proxy_set_header Upgrade $http_upgrade;
+    #         proxy_set_header Connection "upgrade";
+    #         proxy_set_header Host $host;
+    #         proxy_set_header X-Real-IP $remote_addr;
+    #         proxy_set_header X-Forwarded-For $proxy_add_x_forwarded_for;
+    #         proxy_set_header X-Forwarded-Proto $scheme;
+    #
+    #         # Cache static assets
+    #         location ~* \.(jpg|jpeg|png|gif|ico|css|js|svg|woff|woff2|ttf|eot)$ {
+    #             proxy_pass http://web_backend;
+    #             proxy_cache cache;
+    #             proxy_cache_valid 200 30d;
+    #             add_header Cache-Control "public, immutable";
+    #             expires 30d;
+    #         }
+    #     }
+    # }
+
+    # Development server (remove in production)
+    server {
+        listen 80;
+        server_name localhost;
+
+        # API Gateway
+        location /api/ {
+            proxy_pass http://api_backend/;
+            proxy_http_version 1.1;
+            proxy_set_header Host $host;
+            proxy_set_header X-Real-IP $remote_addr;
+            proxy_set_header X-Forwarded-For $proxy_add_x_forwarded_for;
+            proxy_set_header X-Forwarded-Proto $scheme;
+        }
+
+        # Web Frontend
+        location / {
+            proxy_pass http://web_backend;
+            proxy_http_version 1.1;
+            proxy_set_header Upgrade $http_upgrade;
+            proxy_set_header Connection "upgrade";
+            proxy_set_header Host $host;
+            proxy_set_header X-Real-IP $remote_addr;
+            proxy_set_header X-Forwarded-For $proxy_add_x_forwarded_for;
+            proxy_set_header X-Forwarded-Proto $scheme;
+        }
+    }
+}
