+++ services/api-gateway/.env.example.bak
@@ -1,0 +1,23 @@
+# Example environment variables for API Gateway
+# Copy this file to .env.development for local development
+
+# Airtable Configuration (Required)
+AIRTABLE_PAT=pat_your_actual_token_here  # Personal Access Token from Airtable
+AIRTABLE_BASE_ID=app_your_actual_base_id_here  # Base ID from Airtable
+
+# Redis Configuration
+REDIS_URL=redis://localhost:6379
+
+# API Configuration
+ENVIRONMENT=development
+LOG_LEVEL=DEBUG
+PORT=8080
+
+# JWT Configuration (Development - change in production)
+JWT_SECRET_KEY=dev-secret-key-change-in-production
+JWT_ALGORITHM=HS256
+JWT_ACCESS_TOKEN_EXPIRE_MINUTES=60
+
+# Rate Limiting
+RATE_LIMIT_PER_MINUTE=120
+RATE_LIMIT_PER_HOUR=5000
