+++ services/api-gateway/src/batch/__init__.py
@@ -1,19 +1,20 @@
 """Batch processing module for API Gateway."""
 
-from .member_tasks import MemberTaskManager
-from .task_queue import (
-    BatchProcessor,
-    TaskPriority,
-    TaskQueue,
-    batch_processor,
-    task_queue,
-)
+# Temporarily disabled imports for Docker testing
+# from .member_tasks import MemberTaskManager
+# from .task_queue import (
+#     BatchProcessor,
+#     TaskPriority,
+#     TaskQueue,
+#     batch_processor,
+#     task_queue,
+# )
 
 __all__ = [
-    "TaskQueue",
-    "BatchProcessor",
-    "TaskPriority",
-    "task_queue",
-    "batch_processor",
-    "MemberTaskManager",
+    # "TaskQueue",
+    # "BatchProcessor", 
+    # "TaskPriority",
+    # "task_queue",
+    # "batch_processor",
+    # "MemberTaskManager",
 ]
