+++ services/data-processor/.env.example
@@ -1,0 +1,26 @@
+# Data Processor Service Environment Variables
+# ============================================
+
+# Service Configuration
+ENVIRONMENT=development
+SERVICE_NAME=data-processor
+LOG_LEVEL=INFO
+
+# Database
+DATABASE_URL=postgresql://seiji_watch_user:seiji_watch_pass@postgres:5432/seiji_watch
+
+# Redis Queue
+REDIS_URL=redis://redis:6379
+
+# Airtable
+AIRTABLE_ACCESS_TOKEN=pat.xxxxxxxxxxxxx
+AIRTABLE_BASE_ID=appXXXXXXXXXXXXXX
+
+# OpenAI
+OPENAI_API_KEY=sk-xxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxx
+OPENAI_MODEL=gpt-4-turbo-preview
+
+# Processing Configuration
+BATCH_SIZE=50
+MAX_CONCURRENT_JOBS=5
+PROCESSING_TIMEOUT_SECONDS=300
