+++ services/data-processor/pyproject.toml
@@ -26,6 +26,7 @@ pytest = "^7.4.0"
 pytest-asyncio = "^0.21.0"
 ruff = "^0.7.0"
 types-requests = "^2.31.0"
+watchdog = {extras = ["watchmedo"], version = "^3.0.0"}
 
 [build-system]
 requires = ["poetry-core"]
