+++ services/diet-scraper/.dockerignore
@@ -46,7 +46,6 @@ desktop.ini
 README.md
 docs/
 *.md
-reports/
 
 # Test files
 tests/
@@ -86,18 +85,14 @@ logs/
 *.backup
 tmp/
 temp/
-*.wav
-*.mp3
-*.mp4
 
 # Scripts (development only)
 scripts/
-*.py.bak
 
 # Data files (should be mounted as volumes)
 data/
-processed_data/
-exports/
+scraped_data/
+downloads/
 
 # Coverage reports
 coverage/
@@ -106,9 +101,3 @@ coverage/
 # Profiling data
 *.prof
 *.lprof
-
-# Analysis scripts (development only)
-*_analysis.py
-*_analyzer.py
-*_inspection.py
-*_diagnosis.py
