+++ services/diet-scraper/.env.example
@@ -1,0 +1,26 @@
+# Diet Scraper Service Environment Variables
+# ==========================================
+
+# Service Configuration
+ENVIRONMENT=development
+SERVICE_NAME=diet-scraper
+LOG_LEVEL=INFO
+
+# Database
+DATABASE_URL=postgresql://seiji_watch_user:seiji_watch_pass@postgres:5432/seiji_watch
+
+# Redis Queue
+REDIS_URL=redis://redis:6379
+
+# Airtable
+AIRTABLE_ACCESS_TOKEN=pat.xxxxxxxxxxxxx
+AIRTABLE_BASE_ID=appXXXXXXXXXXXXXX
+
+# Scraping Configuration
+SCRAPING_DELAY_SECONDS=2
+SCRAPING_USER_AGENT=Diet-Issue-Tracker/1.0
+SCRAPING_MAX_RETRIES=3
+NDL_API_RATE_LIMIT=3
+
+# Storage
+SCRAPED_DATA_PATH=/data/scraped
