+++ services/diet-scraper/pyproject.toml
@@ -1,0 +1,70 @@
+[tool.poetry]
+name = "diet-scraper"
+version = "0.1.0"
+description = "Data collection service from Diet websites and NDL API"
+authors = ["Diet Issue Tracker Team"]
+packages = [{include = "*", from = "src"}]
+
+[tool.poetry.dependencies]
+python = "^3.11"
+requests = "^2.31.0"
+beautifulsoup4 = "^4.12.0"
+scrapy = "^2.11.0"
+pdfplumber = "^0.10.0"
+janome = "^0.4.2"
+aiohttp = "^3.9.0"
+sqlalchemy = "^2.0.0"
+psycopg2-binary = "^2.9.0"
+pydantic = "^2.5.0"
+python-dotenv = "^1.0.0"
+tenacity = "^8.2.0"
+redis = "^5.0.0"
+google-cloud-pubsub = "^2.18.0"
+shared = {path = "../../shared", develop = true}
+
+[tool.poetry.group.dev.dependencies]
+pytest = "^7.4.0"
+pytest-asyncio = "^0.21.0"
+pytest-cov = "^4.1.0"
+black = "^23.12.0"
+ruff = "^0.7.0"
+mypy = "^1.7.0"
+types-requests = "^2.31.0"
+types-beautifulsoup4 = "^4.12.0"
+httpx = "^0.25.0"
+watchdog = {extras = ["watchmedo"], version = "^3.0.0"}
+
+[build-system]
+requires = ["poetry-core"]
+build-backend = "poetry.core.masonry.api"
+
+[tool.ruff]
+line-length = 100
+target-version = "py311"
+select = [
+    "E",  # pycodestyle errors
+    "W",  # pycodestyle warnings
+    "F",  # pyflakes
+    "I",  # isort
+    "B",  # flake8-bugbear
+    "C4", # flake8-comprehensions
+    "UP", # pyupgrade
+]
+ignore = ["E501", "B008"]
+
+[tool.mypy]
+python_version = "3.11"
+warn_return_any = true
+warn_unused_configs = true
+disallow_untyped_defs = true
+disallow_any_unimported = false
+no_implicit_optional = true
+check_untyped_defs = true
+namespace_packages = true
+ignore_missing_imports = true
+
+[tool.pytest.ini_options]
+testpaths = ["tests"]
+python_files = ["test_*.py", "*_test.py"]
+addopts = "-v --tb=short --strict-markers"
+asyncio_mode = "auto"
