+++ services/notifications-worker/.env.example
@@ -1,0 +1,26 @@
+# Notifications Worker Service Environment Variables
+# ==================================================
+
+# Service Configuration
+ENVIRONMENT=development
+SERVICE_NAME=notifications-worker
+LOG_LEVEL=INFO
+
+# Database
+DATABASE_URL=postgresql://seiji_watch_user:seiji_watch_pass@postgres:5432/seiji_watch
+
+# Redis Queue
+REDIS_URL=redis://redis:6379
+
+# SendGrid Email
+SENDGRID_API_KEY=SG.xxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxx
+SENDGRID_FROM_EMAIL=notifications@seiji-watch.com
+SENDGRID_FROM_NAME=Diet Issue Tracker
+
+# Discord (Optional)
+DISCORD_WEBHOOK_URL=https://discord.com/api/webhooks/xxxx/xxxx
+
+# Notification Settings
+DAILY_SUMMARY_TIME=22:00  # JST
+MAX_EMAILS_PER_BATCH=100
+RATE_LIMIT_PER_MINUTE=100
