+++ services/stt-worker/.env.example
@@ -1,0 +1,29 @@
+# STT Worker Service Environment Variables
+# ========================================
+
+# Service Configuration
+ENVIRONMENT=development
+SERVICE_NAME=stt-worker
+LOG_LEVEL=INFO
+
+# Database
+DATABASE_URL=postgresql://seiji_watch_user:seiji_watch_pass@postgres:5432/seiji_watch
+
+# Redis Queue
+REDIS_URL=redis://redis:6379
+
+# OpenAI
+OPENAI_API_KEY=sk-xxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxx
+
+# Whisper Configuration
+WHISPER_MODEL=large-v3
+WHISPER_LANGUAGE=ja
+WHISPER_DEVICE=cpu
+
+# YouTube Download
+YT_DLP_QUALITY=720p
+
+# Storage
+AUDIO_DATA_PATH=/data/audio
+TRANSCRIPTS_DATA_PATH=/data/transcripts
+CLOUD_STORAGE_BUCKET=seiji-watch-storage
