+++ services/stt-worker/Dockerfile
@@ -1,0 +1,197 @@
+# Multi-stage build for STT Worker service
+# Optimized for layer caching, security, and minimal image size
+
+# ============ Stage 1: Builder ============
+FROM python:3.11-slim as builder
+
+# Build arguments for flexibility
+ARG POETRY_VERSION=1.7.1
+ARG PIP_VERSION=24.0
+
+# Set environment variables for build
+ENV PYTHONUNBUFFERED=1 \
+    PYTHONDONTWRITEBYTECODE=1 \
+    PIP_NO_CACHE_DIR=1 \
+    PIP_DISABLE_PIP_VERSION_CHECK=1 \
+    POETRY_NO_INTERACTION=1 \
+    POETRY_VIRTUALENVS_IN_PROJECT=1 \
+    POETRY_VIRTUALENVS_CREATE=1 \
+    POETRY_CACHE_DIR=/tmp/poetry_cache
+
+# Install system dependencies for building (single layer)
+RUN apt-get update && apt-get install -y --no-install-recommends \
+    build-essential \
+    curl \
+    gcc \
+    g++ \
+    && apt-get clean \
+    && rm -rf /var/lib/apt/lists/*
+
+# Upgrade pip and install Poetry
+RUN pip install --upgrade pip==${PIP_VERSION} && \
+    pip install poetry==${POETRY_VERSION}
+
+WORKDIR /app
+
+# Copy only dependency files first (better caching)
+COPY pyproject.toml poetry.lock* ./
+
+# Install dependencies (without dev dependencies)
+RUN poetry install --without dev --no-root && \
+    rm -rf ${POETRY_CACHE_DIR}
+
+# ============ Stage 2: Development ============
+FROM python:3.11-slim as development
+
+# Set environment variables
+ENV PYTHONUNBUFFERED=1 \
+    PYTHONDONTWRITEBYTECODE=1 \
+    VIRTUAL_ENV=/app/.venv \
+    PATH="/app/.venv/bin:$PATH"
+
+# Install runtime + dev dependencies
+RUN apt-get update && apt-get install -y --no-install-recommends \
+    libpq-dev \
+    curl \
+    git \
+    ffmpeg \
+    wget \
+    && apt-get clean \
+    && rm -rf /var/lib/apt/lists/*
+
+WORKDIR /app
+
+# Copy virtual environment from builder
+COPY --from=builder /app/.venv /app/.venv
+
+# Copy application code
+COPY . .
+
+# Install dev dependencies
+RUN pip install --no-cache-dir \
+    pytest \
+    pytest-cov \
+    pytest-asyncio \
+    black \
+    ruff \
+    mypy \
+    ipdb
+
+# Create directories for audio processing
+RUN mkdir -p /app/audio /app/transcripts /app/logs /app/models
+
+# Health check endpoint port
+EXPOSE 8083
+
+# Health check
+HEALTHCHECK --interval=30s --timeout=10s --start-period=5s --retries=3 \
+    CMD curl -f http://localhost:8083/health || exit 1
+
+# Development mode
+CMD ["python", "-m", "src.main", "--dev"]
+
+# ============ Stage 3: Testing ============
+FROM python:3.11-slim as test
+
+# Set environment variables
+ENV PYTHONUNBUFFERED=1 \
+    PYTHONDONTWRITEBYTECODE=1 \
+    VIRTUAL_ENV=/app/.venv \
+    PATH="/app/.venv/bin:$PATH"
+
+# Install runtime dependencies
+RUN apt-get update && apt-get install -y --no-install-recommends \
+    libpq-dev \
+    curl \
+    ffmpeg \
+    && apt-get clean \
+    && rm -rf /var/lib/apt/lists/*
+
+WORKDIR /app
+
+# Copy virtual environment from builder
+COPY --from=builder /app/.venv /app/.venv
+
+# Copy application code and tests
+COPY . .
+
+# Install test dependencies
+RUN pip install --no-cache-dir \
+    pytest \
+    pytest-cov \
+    pytest-asyncio \
+    pytest-mock \
+    responses
+
+# Run tests by default
+CMD ["pytest", "-v", "--cov=src", "--cov-report=term-missing"]
+
+# ============ Stage 4: Production ============
+FROM python:3.11-slim as production
+
+# Build arguments
+ARG GIT_COMMIT=unknown
+ARG BUILD_DATE=unknown
+ARG VERSION=latest
+
+# Labels for metadata
+LABEL org.opencontainers.image.title="STT Worker" \
+      org.opencontainers.image.description="Diet Issue Tracker Speech-to-Text Worker Service" \
+      org.opencontainers.image.version="${VERSION}" \
+      org.opencontainers.image.created="${BUILD_DATE}" \
+      org.opencontainers.image.revision="${GIT_COMMIT}" \
+      org.opencontainers.image.source="https://github.com/your-repo/seiji-watch"
+
+# Set environment variables
+ENV PYTHONUNBUFFERED=1 \
+    PYTHONDONTWRITEBYTECODE=1 \
+    PYTHONHASHSEED=random \
+    VIRTUAL_ENV=/app/.venv \
+    PATH="/app/.venv/bin:$PATH" \
+    # Security
+    PYTHONIOENCODING=utf-8 \
+    LANG=C.UTF-8 \
+    LC_ALL=C.UTF-8 \
+    # Whisper settings
+    OMP_NUM_THREADS=4
+
+# Install runtime dependencies only (minimal)
+# Include ffmpeg for audio processing
+RUN apt-get update && apt-get install -y --no-install-recommends \
+    libpq-dev \
+    curl \
+    ffmpeg \
+    ca-certificates \
+    && apt-get clean \
+    && rm -rf /var/lib/apt/lists/* \
+    && rm -rf /tmp/* /var/tmp/*
+
+# Create non-root user with specific UID/GID
+RUN groupadd -r -g 1000 appuser && \
+    useradd -r -u 1000 -g appuser -d /app -s /sbin/nologin appuser
+
+WORKDIR /app
+
+# Copy virtual environment from builder (as root for proper permissions)
+COPY --from=builder --chown=appuser:appuser /app/.venv /app/.venv
+
+# Copy only necessary application files (exclude tests, docs, etc.)
+COPY --chown=appuser:appuser src/ ./src/
+COPY --chown=appuser:appuser pyproject.toml ./
+
+# Create necessary directories with proper permissions
+RUN mkdir -p /app/audio /app/transcripts /app/logs /app/models && \
+    chown -R appuser:appuser /app/audio /app/transcripts /app/logs /app/models
+
+# Switch to non-root user
+USER appuser
+
+# Health check endpoint
+EXPOSE 8083
+
+# Health check with shorter timeout for production
+HEALTHCHECK --interval=30s --timeout=5s --start-period=10s --retries=3 \
+    CMD curl -f http://localhost:8083/health || exit 1
+
+# Production mode - runs as a worker processing queue messages
+CMD ["python", "-m", "src.main"]
