+++ services/vector-store/.dockerignore
@@ -46,7 +46,6 @@ desktop.ini
 README.md
 docs/
 *.md
-reports/
 
 # Test files
 tests/
@@ -86,18 +85,23 @@ logs/
 *.backup
 tmp/
 temp/
-*.wav
-*.mp3
-*.mp4
+
+# Embeddings cache (should be stored in persistent storage)
+embeddings/
+*.npy
+*.npz
+*.pkl
+*.pickle
+
+# Model files (should be downloaded at runtime)
+models/
+*.pt
+*.pth
+*.onnx
+*.bin
 
 # Scripts (development only)
 scripts/
-*.py.bak
-
-# Data files (should be mounted as volumes)
-data/
-processed_data/
-exports/
 
 # Coverage reports
 coverage/
@@ -107,8 +111,6 @@ coverage/
 *.prof
 *.lprof
 
-# Analysis scripts (development only)
-*_analysis.py
-*_analyzer.py
-*_inspection.py
-*_diagnosis.py
+# Database dumps
+*.sql
+*.dump
