+++ services/vector-store/.env.example
@@ -1,0 +1,26 @@
+# Vector Store Service Environment Variables
+# ==========================================
+
+# Service Configuration
+ENVIRONMENT=development
+SERVICE_NAME=vector-store
+LOG_LEVEL=INFO
+
+# Database
+DATABASE_URL=postgresql://seiji_watch_user:seiji_watch_pass@postgres:5432/seiji_watch
+
+# Redis Queue
+REDIS_URL=redis://redis:6379
+
+# Weaviate
+WEAVIATE_URL=https://your-cluster.weaviate.network
+WEAVIATE_API_KEY=your-weaviate-api-key
+
+# OpenAI
+OPENAI_API_KEY=sk-xxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxx
+OPENAI_EMBEDDING_MODEL=text-embedding-3-large
+
+# Embedding Configuration
+BATCH_SIZE=100
+MAX_TOKENS=8192
+EMBEDDING_DIMENSION=1536
