+++ services/web-frontend/.dockerignore
@@ -1,40 +1,27 @@
 # Git files
 .git
 .gitignore
+.gitattributes
 
 # Dependencies
-node_modules
+node_modules/
 npm-debug.log*
 yarn-debug.log*
 yarn-error.log*
+.pnpm-debug.log*
 
-# Next.js
+# Next.js build output
 .next/
 out/
-
-# Production build
 build/
+dist/
 
-# Environment variables
-.env*.local
-.env.development
-.env.test
-
-# IDE files
-.vscode
-.idea
-*.swp
-*.swo
-
-# OS files
-.DS_Store
-Thumbs.db
-
-# Documentation
-README.md
-docs/
-
-# Test files
+# Testing
+coverage/
+.nyc_output/
+playwright-report/
+test-results/
+jest-coverage/
 **/*.test.js
 **/*.test.ts
 **/*.test.tsx
@@ -42,20 +29,83 @@ docs/
 **/*.spec.ts
 **/*.spec.tsx
 __tests__/
-coverage/
+
+# Environment files
+.env
+.env.local
+.env.development
+.env.test
+.env.production
+*.env.example
+
+# IDE files
+.vscode/
+.idea/
+*.swp
+*.swo
+*~
+.project
+.settings/
+
+# OS files
+.DS_Store
+Thumbs.db
+desktop.ini
+
+# Documentation
+README.md
+docs/
+*.md
+
+# Development files
+.eslintrc*
+.prettierrc*
+.stylelintrc*
+jest.config.*
+jest.setup.*
+tsconfig.tsbuildinfo
+playwright.config.*
+.eslintcache
+
+# CI/CD files
+.github/
+.gitlab-ci.yml
+.travis.yml
+cloudbuild.yaml
+Dockerfile*
+docker-compose*.yml
+
+# Logs
+*.log
+logs/
+*.log.*
+dev.log
+frontend.log
+
+# Temporary files
+*.tmp
+*.bak
+*.backup
+tmp/
+temp/
+
+# Scripts (development only)
+scripts/
+
+# Config backup
+config/backup/
+
+# TypeScript cache
+*.tsbuildinfo
+.tscache/
 
 # Storybook
 .storybook/
 storybook-static/
 
-# Logs
-*.log
-logs/
-
-# Cache
-.eslintcache
-.next/cache/
-
 # Misc
+.vercel
+.turbo
+.cache/
 *.tgz
 *.tar.gz
